feat:编写新增商品部分代码

+++ product/src/main/java/com/sxx/product/vo/Images.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Data
 public class Images {
 
-  private String imgUrl;
-  private int defaultImg;
+    private String imgUrl;
+    private Byte defaultImg;
 
 }
